Adding the Table of monthly climate data
This Table contains the monthly raw climate data from ANUClimate and TerraClimate (see Methods in the paper), as well as the 30-year climatology and monthly standardised anomalies (see Methods, and Supplementary R code). Because the growth observations were available at a time resolution of 2 to 5 years, the monthly values were averaged over the months encompassed by the corresponding growth periods to generate the climate covariates of the growth models (see Climate data and Data analysis in the Methods section of the paper).